moved Types.js to business
