+++ src/business/User.js
@@ -3,7 +3,7 @@
 
 import * as dbRef from '../constants/dbRef';
 import { db } from '../data/firebase/firebase';
-import type { KeyType } from '../Types';
+import type { KeyType } from './Types';
 
 export const doCreateUser = (
   userId: KeyType, name: string, email: string, imageUrl: ?string,
