+++ src/components/MyPins.js
@@ -5,7 +5,7 @@ import { Route } from 'react-router-dom';
 import * as routes from '../constants/routes';
 import ListPins from './Pin/ListPins';
 import PinDetails from './Pin/PinDetails';
-import type { AuthUserType } from '../Types';
+import type { AuthUserType } from '../business/Types';
 
 type Props = {
   authUser: AuthUserType
